+++ CHANGES.md
@@ -1,4 +1,7 @@
 ## Version 0.11.1:
+ * [sertop]  Set default value of allow-sprop to be true in agreement with upstream coq v8.11
+             and added option '--disallow-sprop' to optionally switch it off
+			 (--disallow-sprop forbids using the proof irrelevant SProp sort)
 
  * [sertop]  Added option `--topfile` to sertop to set top name from
              a filename